backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.7
botocore==1.35.7
chardet==5.2.0
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
jmespath==1.0.1
openpyxl==3.1.5
polars==1.6.0
pydantic==2.8.2
pydantic-core==2.20.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing-extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.6
